% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSplinePredDeriv.R
\name{cSplinePredDeriv}
\alias{cSplinePredDeriv}
\title{Calculate the derivatives of coefficients of a fixed effect regression spline (convexity).}
\usage{
cSplinePredDeriv(t, theta, m)
}
\arguments{
\item{t}{The knot sequence vector in exposure-outcome model.}

\item{theta}{The coefficient vector of C-spline bases in exposure-outcome model.}

\item{m}{The mediator value within min(t) and max(t).}
}
\value{
The derivatives of beta.
}
\description{
Calculate the derivatives of coefficients of a fixed effect regression spline (convexity).
}
