% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionAsymp.R
\name{distributionAsymp}
\alias{distributionAsymp}
\title{Asymptotic null distribution of the test statistic}
\usage{
distributionAsymp(
  nIter,
  nBasis,
  sigmai = NULL,
  sigmac = NULL,
  testStatIncr = NULL,
  testStatDecr = NULL,
  testStatConv = NULL,
  testStatConc = NULL
)
}
\arguments{
\item{nIter}{The number of iterations.}

\item{nBasis}{The number of bases.}

\item{sigmai}{The covariance matrix of coefficient estimates of I-Spline bases (monotonicity)}

\item{sigmac}{The covariance matrix of coefficient estimates of C-Spline bases (convexity)}

\item{testStatIncr}{The test statistic generated from iSplineFER/iSpineMER}

\item{testStatDecr}{The test statistic generated from iSplineFER/iSpineMER}

\item{testStatConv}{The test statistic generated from cSplineFER/cSpineMER}

\item{testStatConc}{The test statistic generated from cSplineFER/cSpineMER}
}
\value{
A list of weights of beta distribution and p-values.
}
\description{
Asymptotic null distribution of the test statistic
}
