% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSplineExpDeriv.R
\name{iSplineExpDeriv}
\alias{iSplineExpDeriv}
\title{Calculate the derivatives of coefficients of a fixed effect regression spline (monotonicity).}
\usage{
iSplineExpDeriv(t, theta, sigma2, gamma0, gamma1, gamma2, a, c)
}
\arguments{
\item{t}{The knot sequence vector in exposure-outcome model.}

\item{theta}{The coefficient vector of I-spline bases in exposure-outcome model.}

\item{sigma2}{The residual standard deviation of exposure-mediator model.}

\item{gamma0}{The coefficient of intercept of exposure-mediator model.}

\item{gamma1}{The coefficient of exposure of exposure-mediator model.}

\item{gamma2}{The coefficient vector of confounders of exposure-mediator model.}

\item{a}{The value of exposure (0/1).}

\item{c}{The values of confounders.}
}
\value{
A list of derivatives of beta, gamma and sigma2^2.
}
\description{
Calculate the derivatives of coefficients of a fixed effect regression spline (monotonicity).
}
