% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSplineFER.R
\name{iSplineFER}
\alias{iSplineFER}
\title{Fit a semi-parametric shape-restricted fixed effect regression spline (monotonicity).}
\usage{
iSplineFER(
  lmFormula,
  dataset,
  varName,
  knotType = 2,
  preKnot = NULL,
  nBasis = 5,
  increasing = T
)
}
\arguments{
\item{lmFormula}{A linear model formula.}

\item{dataset}{A data frame.}

\item{varName}{The name of the main effect.}

\item{knotType}{The knot type: 1=equal-spaced, 2=quantile, 3=pre-specified.}

\item{preKnot}{The pre-specified knots.}

\item{nBasis}{The number of bases.}

\item{increasing}{Increasing shape or Decreasing shape: T=Increasing, F=Decreasing}
}
\value{
A list of knots, data set, lm model, summary of lm model, coefficient estimates of bases, covariance of coefficient estimates of bases, combined coefficient estimates, test statistic.
}
\description{
Fit a semi-parametric shape-restricted fixed effect regression spline (monotonicity).
}
