% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSplineGen.R
\name{iSplineGen}
\alias{iSplineGen}
\title{Generate I-Spline Basis}
\usage{
iSplineGen(
  dataset,
  varName,
  knotType = 1,
  preKnot = NULL,
  nBasis = 5,
  plot = F
)
}
\arguments{
\item{dataset}{A data frame.}

\item{varName}{The name of the main effect.}

\item{knotType}{The knot type: 1=equal-spaced, 2=quantile, 3=pre-specified.}

\item{preKnot}{The pre-specified knots.}

\item{nBasis}{The number of bases.}

\item{plot}{Plot the basis function or not: T=yes, F=no.}
}
\value{
A list of knots, bases, data set.
}
\description{
Generate I-Spline Basis
}
