% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSplinePred.R
\name{iSplinePred}
\alias{iSplinePred}
\title{Calculate the predicted value of a fixed effect regression spline (monotonicity).}
\usage{
iSplinePred(t, theta, m)
}
\arguments{
\item{t}{The knot sequence vector in exposure-outcome model.}

\item{theta}{The coefficient vector of I-spline bases in exposure-outcome model.}

\item{m}{The mediator value within min(t) and max(t).}
}
\value{
The predicted value of a fixed effect regression spline.
}
\description{
Calculate the predicted value of a fixed effect regression spline (monotonicity).
}
