% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ySimGen.R
\name{ySimGen}
\alias{ySimGen}
\title{Generate y using existed X, coefficients of X, variance of random effect and variance of noise}
\usage{
ySimGen(
  dataset,
  seed,
  coefVec,
  fixMat,
  randomVar = NULL,
  randomSD = NULL,
  noiseSD
)
}
\arguments{
\item{dataset}{A data frame.}

\item{seed}{The seed.}

\item{coefVec}{The coefficient vector of the fixed effect matrix.}

\item{fixMat}{The fixed effect matrix.}

\item{randomVar}{The random effect variable in the data set.}

\item{randomSD}{The SD of random effect.}

\item{noiseSD}{The SD of noise.}
}
\value{
A data set with generated y.
}
\description{
Generate y using existed X, coefficients of X, variance of random effect and variance of noise
}
